addressed changes from PR
