+++ darwin/client.py
@@ -12,12 +12,9 @@ from darwin.exceptions import (
     InsufficientStorage,
     InvalidLogin,
     MissingConfig,
-    NameTaken,
     NotFound,
-    Unauthenticated,
     Unauthorized,
-    ValidationError,
-)
+    name_taken, validation_error)
 from darwin.utils import is_project_dir, urljoin
 
 
@@ -434,11 +431,3 @@ class Client:
         return f"Client(default_team={self.default_team})"
 
 
-def name_taken(code, body):
-    if code == 422 and body["errors"]["name"][0] == "has already been taken":
-        raise NameTaken
-
-
-def validation_error(code, body):
-    if code == 422:
-        raise ValidationError(body)
